refactor publicationservice class
