+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/service/artefact/ArtefactTriggerServiceTest.java
@@ -1,0 +1,329 @@
+package uk.gov.hmcts.reform.pip.data.management.service.artefact;
+
+import nl.altindag.log.LogCaptor;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.reform.pip.data.management.database.ArtefactRepository;
+import uk.gov.hmcts.reform.pip.data.management.database.LocationRepository;
+import uk.gov.hmcts.reform.pip.data.management.models.location.Location;
+import uk.gov.hmcts.reform.pip.data.management.models.location.LocationCsv;
+import uk.gov.hmcts.reform.pip.data.management.models.location.LocationType;
+import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
+import uk.gov.hmcts.reform.pip.data.management.models.publication.Language;
+import uk.gov.hmcts.reform.pip.data.management.models.publication.ListType;
+import uk.gov.hmcts.reform.pip.data.management.models.publication.Sensitivity;
+import uk.gov.hmcts.reform.pip.data.management.service.PublicationServicesService;
+import uk.gov.hmcts.reform.pip.data.management.service.SubscriptionManagementService;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.lenient;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.reform.pip.data.management.helpers.ConstantsTestHelper.MESSAGES_MATCH;
+
+@ExtendWith(MockitoExtension.class)
+class ArtefactTriggerServiceTest {
+
+    @Mock
+    ArtefactRepository artefactRepository;
+
+    @Mock
+    LocationRepository locationRepository;
+
+    @Mock
+    SubscriptionManagementService subscriptionManagementService;
+
+    @Mock
+    PublicationServicesService publicationServicesService;
+
+    @InjectMocks
+    ArtefactTriggerService artefactTriggerService;
+
+    private static final UUID ARTEFACT_ID = UUID.randomUUID();
+    private static final String SOURCE_ARTEFACT_ID = "1234";
+    private static final String PROVENANCE = "provenance";
+    private static final String PROVENANCE_ID = "1234";
+    private static final String PAYLOAD_URL = "https://ThisIsATestPayload";
+    private static final String LOCATION_ID = "123";
+    private static final String TEST_KEY = "TestKey";
+    private static final String TEST_VALUE = "TestValue";
+    private static final Map<String, List<Object>> SEARCH_VALUES = new ConcurrentHashMap<>();
+    private static final String SUCCESSFUL_TRIGGER = "success - subscription sent";
+    private static final String SUCCESS = "Success";
+
+    private static final String NO_COURT_EXISTS_IN_REFERENCE_DATA = "NoMatch1234";
+
+    private static final LocalDateTime CONTENT_DATE = LocalDateTime.now();
+    private static final LocalDateTime START_OF_TODAY_CONTENT_DATE = LocalDateTime.now().toLocalDate().atStartOfDay();
+
+    private Artefact artefact;
+    private Artefact artefactWithPayloadUrl;
+    private Artefact artefactWithIdAndPayloadUrl;
+    private Artefact artefactInTheFuture;
+    private Artefact artefactFromThePast;
+    private Artefact artefactFromNow;
+    private Artefact artefactWithNullDateTo;
+    private Artefact artefactWithSameDateFromAndTo;
+    private Artefact noMatchArtefact;
+
+    private Location location;
+
+    @BeforeAll
+    public static void setupSearchValues() {
+        SEARCH_VALUES.put(TEST_KEY, List.of(TEST_VALUE));
+    }
+
+    @BeforeEach
+    void setup() {
+        createPayloads();
+        createClassifiedPayloads();
+
+        initialiseCourts();
+
+        lenient().when(artefactRepository.findArtefactByUpdateLogic(artefact.getLocationId(),artefact.getContentDate(),
+                                                                    artefact.getLanguage().name(),
+                                                                    artefact.getListType().name(),
+                                                                    artefact.getProvenance()))
+            .thenReturn(Optional.empty());
+        lenient().when(artefactRepository.save(artefactWithPayloadUrl)).thenReturn(artefactWithIdAndPayloadUrl);
+        lenient().when(locationRepository.findByLocationIdByProvenance(PROVENANCE, PROVENANCE_ID,
+                                                                       LocationType.VENUE.name()))
+            .thenReturn(Optional.of(location));
+    }
+
+    private void createPayloads() {
+        artefact = Artefact.builder()
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .locationId(PROVENANCE_ID)
+            .contentDate(START_OF_TODAY_CONTENT_DATE)
+            .listType(ListType.CIVIL_DAILY_CAUSE_LIST)
+            .language(Language.ENGLISH)
+            .sensitivity(Sensitivity.PUBLIC)
+            .build();
+
+        artefactWithPayloadUrl = Artefact.builder()
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .locationId(LOCATION_ID)
+            .contentDate(START_OF_TODAY_CONTENT_DATE)
+            .listType(ListType.CIVIL_DAILY_CAUSE_LIST)
+            .language(Language.ENGLISH)
+            .sensitivity(Sensitivity.PUBLIC)
+            .build();
+
+        artefactWithIdAndPayloadUrl = Artefact.builder()
+            .artefactId(ARTEFACT_ID)
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .locationId(LOCATION_ID)
+            .contentDate(CONTENT_DATE)
+            .listType(ListType.CIVIL_DAILY_CAUSE_LIST)
+            .language(Language.ENGLISH)
+            .sensitivity(Sensitivity.PUBLIC)
+            .build();
+
+        artefactFromThePast = Artefact.builder()
+            .artefactId(ARTEFACT_ID)
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .displayFrom(LocalDateTime.now().minusDays(1))
+            .displayTo(LocalDateTime.now().plusDays(1))
+            .build();
+
+        artefactFromNow = Artefact.builder()
+            .artefactId(ARTEFACT_ID)
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .displayFrom(LocalDateTime.now())
+            .displayTo(LocalDateTime.now().plusHours(3))
+            .build();
+
+        artefactWithNullDateTo = Artefact.builder()
+            .artefactId(ARTEFACT_ID)
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .displayFrom(LocalDateTime.now())
+            .displayTo(null)
+            .build();
+
+        artefactWithSameDateFromAndTo = Artefact.builder()
+            .artefactId(ARTEFACT_ID)
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .displayFrom(LocalDateTime.now())
+            .displayTo(LocalDateTime.now())
+            .build();
+
+        artefactInTheFuture = Artefact.builder()
+            .artefactId(ARTEFACT_ID)
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .displayFrom(LocalDateTime.now().plusDays(1))
+            .displayTo(LocalDateTime.now().plusDays(2))
+            .build();
+
+        noMatchArtefact = Artefact.builder()
+            .sourceArtefactId(SOURCE_ARTEFACT_ID)
+            .provenance(PROVENANCE)
+            .payload(PAYLOAD_URL)
+            .search(SEARCH_VALUES)
+            .locationId(NO_COURT_EXISTS_IN_REFERENCE_DATA)
+            .contentDate(CONTENT_DATE)
+            .listType(ListType.CIVIL_DAILY_CAUSE_LIST)
+            .language(Language.ENGLISH)
+            .sensitivity(Sensitivity.PUBLIC)
+            .build();
+    }
+
+    private void createClassifiedPayloads() {
+
+        initialiseCourts();
+
+        lenient().when(artefactRepository.findArtefactByUpdateLogic(artefact.getLocationId(),artefact.getContentDate(),
+                                                                    artefact.getLanguage().name(),
+                                                                    artefact.getListType().name(),
+                                                                    artefact.getProvenance()))
+            .thenReturn(Optional.empty());
+        lenient().when(artefactRepository.save(artefactWithPayloadUrl)).thenReturn(artefactWithIdAndPayloadUrl);
+        lenient().when(locationRepository.findByLocationIdByProvenance(PROVENANCE, PROVENANCE_ID,
+                                                                       LocationType.VENUE.name()))
+            .thenReturn(Optional.of(location));
+    }
+
+    private void initialiseCourts() {
+        LocationCsv locationCsvFirstExample = new LocationCsv();
+        locationCsvFirstExample.setLocationName("Court Name First Example");
+        locationCsvFirstExample.setProvenanceLocationType("venue");
+        location = new Location(locationCsvFirstExample);
+        location.setLocationId(1234);
+
+    }
+
+    @Test
+    void testTriggerIfDateIsFuture() throws IOException {
+        try (LogCaptor logCaptor = LogCaptor.forClass(ArtefactTriggerService.class)) {
+            artefactTriggerService.checkAndTriggerSubscriptionManagement(artefactInTheFuture);
+            assertEquals(
+                0,
+                logCaptor.getInfoLogs().size(),
+                "Should not have returned a log as no trigger was sent."
+            );
+        } catch (Exception ex) {
+            throw new IOException(ex.getMessage());
+        }
+    }
+
+    @Test
+    void testTriggerIfDateIsNow() {
+        when(subscriptionManagementService.sendArtefactForSubscription(artefactFromNow)).thenReturn(SUCCESSFUL_TRIGGER);
+        try (LogCaptor logCaptor = LogCaptor.forClass(ArtefactTriggerService.class)) {
+            artefactTriggerService.checkAndTriggerSubscriptionManagement(artefactFromNow);
+            assertEquals(SUCCESSFUL_TRIGGER, logCaptor.getInfoLogs().get(0),
+                         "should have returned the Subscription List."
+            );
+        }
+    }
+
+    @Test
+    void testTriggerIfDateIsPast() {
+        when(subscriptionManagementService.sendArtefactForSubscription(artefactFromThePast)).thenReturn(
+            SUCCESSFUL_TRIGGER);
+        try (LogCaptor logCaptor = LogCaptor.forClass(ArtefactTriggerService.class)) {
+            artefactTriggerService.checkAndTriggerSubscriptionManagement(artefactFromThePast);
+            assertEquals(SUCCESSFUL_TRIGGER, logCaptor.getInfoLogs().get(0),
+                         "Should have returned the subscription list"
+            );
+        }
+    }
+
+    @Test
+    void testTriggerIfDateToNull() {
+        when(subscriptionManagementService.sendArtefactForSubscription(artefactWithNullDateTo)).thenReturn(
+            SUCCESSFUL_TRIGGER);
+        try (LogCaptor logCaptor = LogCaptor.forClass(ArtefactTriggerService.class)) {
+            artefactTriggerService.checkAndTriggerSubscriptionManagement(artefactWithNullDateTo);
+            assertEquals(SUCCESSFUL_TRIGGER, logCaptor.getInfoLogs().get(0),
+                         "Should have returned the subscription list"
+            );
+        }
+    }
+
+    @Test
+    void testTriggerIfSameDateFromTo() {
+        when(subscriptionManagementService.sendArtefactForSubscription(artefactWithSameDateFromAndTo)).thenReturn(
+            SUCCESSFUL_TRIGGER);
+        try (LogCaptor logCaptor = LogCaptor.forClass(ArtefactTriggerService.class)) {
+            artefactTriggerService.checkAndTriggerSubscriptionManagement(artefactWithSameDateFromAndTo);
+            assertEquals(SUCCESSFUL_TRIGGER, logCaptor.getInfoLogs().get(0),
+                         "Should have returned the subscription list"
+            );
+        }
+    }
+
+    @Test
+    void testSendArtefactForSubscription() {
+        when(subscriptionManagementService.sendArtefactForSubscription(artefact))
+            .thenReturn(SUCCESS);
+        assertEquals(SUCCESS, artefactTriggerService.sendArtefactForSubscription(artefact),
+                     MESSAGES_MATCH);
+    }
+
+    @Test
+    void testCheckNewlyActiveArtefactsLogs() throws IOException {
+        try (LogCaptor logCaptor = LogCaptor.forClass(ArtefactTriggerService.class)) {
+            when(artefactRepository.findArtefactsByDisplayFrom(any())).thenReturn(List.of(new Artefact()));
+            when(subscriptionManagementService.sendArtefactForSubscription(any())).thenReturn(SUCCESS);
+            artefactTriggerService.checkNewlyActiveArtefacts();
+            assertEquals(SUCCESS, logCaptor.getInfoLogs().get(0),
+                         MESSAGES_MATCH
+            );
+        } catch (Exception ex) {
+            throw new IOException(ex.getMessage());
+        }
+    }
+
+    @Test
+    void testReportNoMatchArtefacts() {
+        when(artefactRepository.findAllNoMatchArtefacts()).thenReturn(List.of(noMatchArtefact));
+        artefactTriggerService.reportNoMatchArtefacts();
+        verify(publicationServicesService).sendNoMatchArtefactsForReporting(Map.of(PROVENANCE_ID, PROVENANCE));
+    }
+
+    @Test
+    void testReportMatchArtefactsWhenArtefactsNotFound() {
+        when(artefactRepository.findAllNoMatchArtefacts()).thenReturn(Collections.emptyList());
+        artefactTriggerService.reportNoMatchArtefacts();
+        verifyNoInteractions(publicationServicesService);
+    }
+}
